pyproject.toml
[project]
name = "agentic-design-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen>=0.11.5",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
]
